web/backend/requirements.txt
# TerrariaSteamDeckServer Web Backend Dependencies
# Flask-based REST API for server management

# Core framework
Flask==3.0.0
Werkzeug==3.0.1

# CORS support
Flask-CORS==4.0.0

# Authentication
PyJWT==2.8.0

# Environment configuration
python-dotenv==1.0.0

# JSON Schema validation
jsonschema==4.20.0

# WSGI server for production
gunicorn==21.2.0
